% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcTMB-package.R, R/drcTMBtest-package.R
\docType{package}
\name{drcTMB-package}
\alias{drcTMB}
\alias{drcTMB-package}
\title{drcTMB: Fitting dose-response curves using the Template Model Builder}
\description{
Provides a simple user-interface for fitting non-linear mixed-effects models for dose-response analysis.

Provides a simple user-interface for fitting non-linear mixed-effects models for dose-response analysis.
}
\author{
\strong{Maintainer}: Daniel Gerhard \email{00gerhard@gmail.com} (\href{https://orcid.org/0000-0002-9336-3454}{ORCID})

}
\keyword{internal}
