% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmTMBmethods.R
\name{drmTMB-methods}
\alias{drmTMB-methods}
\alias{print.drmTMB}
\alias{predict.drmTMB}
\alias{coef.drmTMB}
\alias{vcov.drmTMB}
\title{drmTMB Methods}
\usage{
\method{print}{drmTMB}(x, ...)

\method{predict}{drmTMB}(object, newdata = NULL, random = TRUE)

\method{coef}{drmTMB}(object, ...)

\method{vcov}{drmTMB}(object, ...)
}
\arguments{
\item{x, object}{a drmTMB object}

\item{...}{anything else}

\item{newdata}{a data frame for prediction. If NULL, the original data is used.}

\item{random}{logical, if TRUE, the random effects are included in the predictions}
}
\value{
Estimated model coefficients, variance-covariance matrix, or predictions
}
\description{
drmTMB Methods
}
