% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmTMB.R
\name{drmTMB}
\alias{drmTMB}
\title{Fit a drmTMB model}
\usage{
drmTMB(
  form,
  fform = NULL,
  rform = NULL,
  data,
  start = NULL,
  fix = NULL,
  family = "gaussian",
  model = "L5"
)
}
\arguments{
\item{form}{formula with name of response and dose scale}

\item{fform}{list with fixed-effect right-hand side formulas for each model parameter}

\item{rform}{random effect formula of the form b1 + b2 + b3 + b4 + b5 ~ 1 | cluster}

\item{data}{a data.frame object}

\item{start}{a vector with starting values for each model parameter. If NA a self-starting function is called.}

\item{fix}{a binary vector with same length as start. If TRUE, a model coefficient is fixed to the starting value.}

\item{family}{Specify the distributional assumptions. One of "gaussian" and "poisson".}

\item{model}{Specify the nonlinear model. "L5".}
}
\value{
A drmTMB object
}
\description{
Fit a drmTMB model
}
