% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmTMBcontrol.R
\name{drmTMBcontrol}
\alias{drmTMBcontrol}
\title{Set control parameters for drcTMB}
\usage{
drmTMBcontrol(optimiser = "nlminb")
}
\arguments{
\item{optimiser}{character string specifying the optimiser to use}
}
\value{
a list with control parameters
}
\description{
Set control parameters for drcTMB
}
