% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edTMBmethods.R
\name{edTMB-methods}
\alias{edTMB-methods}
\alias{print.edTMB}
\alias{coef.edTMB}
\alias{vcov.edTMB}
\title{edTMB Methods}
\usage{
\method{print}{edTMB}(x, ...)

\method{coef}{edTMB}(object, ...)

\method{vcov}{edTMB}(object, ...)
}
\arguments{
\item{x, object}{an edTMB object}

\item{...}{anything else}
}
\value{
Estimated effective dose coefficients
}
\description{
edTMB Methods
}
