% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drmTMBmethods.R
\name{print.drmTMB}
\alias{print.drmTMB}
\title{Title}
\usage{
\method{print}{drmTMB}(x, ...)
}
\arguments{
\item{x}{a drmTMB object}

\item{...}{anything else}
}
\value{
Estimated model coefficients
}
\description{
Title
}
