% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed.R
\name{ed}
\alias{ed}
\title{Effective dose estimation}
\usage{
ed(x, respLev = 50, linfct = diag(5), random = TRUE, residual = FALSE)
}
\arguments{
\item{x}{a drmTMB object}

\item{respLev}{effective dose level, default is 50}

\item{linfct}{a matrix of linear functions to indentify the parameters of the dose-response curve (each row vector results in parameters b1-b5), default is the identity matrix}

\item{random}{logical, if TRUE, the variance-covariance matrix of the random effects is included in the standard error estimation}

\item{residual}{logical, if TRUE, the residual variance is included in the standard error estimation}
}
\value{
a edTMB object
}
\description{
Effective dose estimation
}
