% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed.R
\name{ed}
\alias{ed}
\title{Effective dose estimation}
\usage{
ed(x, respLev)
}
\arguments{
\item{x}{a drmTMB object}

\item{respLev}{effective dose level}
}
\value{
a vector with estimates
}
\description{
Effective dose estimation
}
