% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edTMBmethods.R
\name{edTMB-methods}
\alias{edTMB-methods}
\alias{print.edTMB}
\title{edTMB Methods}
\usage{
\method{print}{edTMB}(x, ...)
}
\arguments{
\item{x}{an edTMB object}

\item{...}{anything else}
}
\value{
Estimated effective dose coefficients
}
\description{
edTMB Methods
}
